feat: Implement backward compatibility for WhatsApp messaging components
- Refactored WhatsAppSender to inherit from the new messaging template sender for backward compatibility.
- Updated message template functions to provide backward compatibility with deprecation warnings.
- Refactored WhatsAppClient to inherit from the new client implementation for backward compatibility.
- Enhanced setup.py with version specifications for dependencies and updated Python version requirement.
- Added unit tests for message handlers and templates to ensure functionality and reliability.
