+++ README.md
@@ -2,17 +2,68 @@
 
 Care IM (Instant Messaging) is a WhatsApp integration plugin for Care, providing a seamless messaging interface through WhatsApp Business API. This plugin enables healthcare facilities to communicate with patients via WhatsApp, supporting automated notifications and interactive messaging.
 
-Key Features:
+## Key Features
+
 - WhatsApp Business API integration
 - Automated patient notifications
 - Two-way messaging support
 - Secure message handling
 - Configurable message templates
+- Modular architecture for easy extension
 
+## Architecture
+
+The plugin is structured in a modular way for better maintainability and extensibility:
+
+- **Core**: Configuration and essential utilities
+- **Messaging**: WhatsApp client, template sending, and message handling
+- **Templates**: Message templates and formatting logic
+- **Handlers**: Specialized handlers for different message types
+- **API**: REST API endpoints for webhook integration
+- **Signals**: Django signal handlers for event-driven messaging
+
+## Usage Examples
+
+### Sending a WhatsApp message
+
+```python
+from care_im.messaging.client import WhatsAppClient
+
+client = WhatsAppClient()
+client.send_message("+919876543210", "Hello from Care IM!")
+```
+
+### Sending a template message
+
+```python
+from care_im.messaging.template_sender import WhatsAppSender
+
+sender = WhatsAppSender()
+sender.send_template(
+    "+919876543210", 
+    "care_appointment_reminder",
+    params={
+        "body": [
+            {"type": "text", "text": "Dr. Smith"},
+            {"type": "text", "text": "March 30, 2025"},
+            {"type": "text", "text": "10:30 AM"}
+        ]
+    }
+)
+```
+
+### Processing an incoming message
+
+```python
+from care_im.messaging.handler import WhatsAppMessageHandler
+
+handler = WhatsAppMessageHandler("+919876543210")
+response = handler.process_message("medications")
+```
 
 ## Local Development
 
-To develop the plug in local environment along with care, follow the steps below:
+To develop the plugin in a local environment along with care, follow the steps below:
 
 1. Go to the care root directory and clone the plugin repository:
 
@@ -80,11 +131,21 @@ plugs = [im_plug]
 
 ## Configuration
 
-The following configurations variables are available for Care Hello:
+The following configuration variables are required for Care IM:
 
-- `HELLO_DUMMY_ENV`: Dummy environment variable for testing
+- `WHATSAPP_ACCESS_TOKEN`: WhatsApp Business API access token
+- `WHATSAPP_PHONE_NUMBER_ID`: Your WhatsApp phone number ID
+- `WHATSAPP_VERIFY_TOKEN`: Token for webhook verification
+- `WHATSAPP_API_VERSION`: WhatsApp API version to use (default: v22.0)
+- `WHATSAPP_BUSINESS_ACCOUNT_ID`: Your WhatsApp Business Account ID
 
-The plugin will try to find the API key from the config first and then from the environment variable.
+## Extending the Plugin
+
+To add a new message handler:
+
+1. Create a new handler class that extends `BaseHandler` in `messaging/handlers/`
+2. Implement your handler logic
+3. Integrate it with the main `WhatsAppMessageHandler` class
 
 ## License
