+++ care_im/__init__.py
@@ -1,0 +1,8 @@
+"""
+Care Instant Messaging (care_im) package.
+
+A WhatsApp integration plugin for Care, providing seamless messaging interface
+through WhatsApp Business API for healthcare facilities.
+"""
+
+__version__ = "0.2.0"
