+++ care_im/core/__init__.py
@@ -1,0 +1,1 @@
+"""Core functionality for the care_im package."""
