+++ care_im/messaging/__init__.py
@@ -1,0 +1,1 @@
+"""WhatsApp messaging functionality for the care_im package."""
