+++ care_im/messaging/handlers/__init__.py
@@ -1,0 +1,1 @@
+"""Specialized message handlers for different types of WhatsApp requests."""
