+++ care_im/signals/__init__.py
@@ -1,1 +1,4 @@
-from .send_message import *
+"""Signal handlers for care_im package."""
+
+# Import signal handlers to register them
+from care_im.signals.send_message import *
