+++ care_im/templates/__init__.py
@@ -1,0 +1,1 @@
+"""Message templates for the care_im package."""
