+++ care_im/utils/send_message_templates.py
@@ -1,180 +1,27 @@
+"""
+Backward compatibility module for WhatsAppSender.
 
+This module provides backward compatibility for code that imports
+WhatsAppSender from the old location.
 
-import requests
-import logging
-from typing import Dict, Any, Optional
-from care_im.settings import settings
+For new code, import from care_im.messaging.template_sender instead.
+"""
 
-logger = logging.getLogger(__name__)
+import warnings
+from care_im.messaging.template_sender import WhatsAppSender as ModernWhatsAppSender
 
-class WhatsAppSender:
+class WhatsAppSender(ModernWhatsAppSender):
     """
-    Class for sending WhatsApp messages using the WhatsApp Business API
+    Backward compatibility wrapper for WhatsAppSender.
+    
+    This class inherits from the new WhatsAppSender for backward compatibility.
     """
-
-    def __init__(self):
-        """
-        Initialize the WhatsApp sender
-        """
-        self.access_token = settings.WHATSAPP_ACCESS_TOKEN
-        self.phone_number_id = settings.WHATSAPP_PHONE_NUMBER_ID
-        self.api_url = f"https://graph.facebook.com/v22.0/{settings.WHATSAPP_PHONE_NUMBER_ID}/messages"
-
-    # def send_template(self, to_number: str, template_name: str, language: str = "en",
-    #                   params: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
-    #     """
-    #     Send a WhatsApp template message
-
-    #     Args:
-    #         to_number: Recipient's phone number with country code (e.g. "911234567890")
-    #         template_name: Name of the template to send
-    #         language: Language code for the template
-    #         params: Parameters to pass to the template (if any)
-
-    #     Returns:
-    #         API response as dictionary
-    #     """
-    #     logger.info(f"Sending WhatsApp template to {to_number} - {template_name}")
-    #     headers = {
-    #         "Authorization": f"Bearer {self.access_token}",
-    #         "Content-Type": "application/json"
-    #     }
-
-    #     # Basic payload for template message
-    #     payload = {
-    #         "messaging_product": "whatsapp",
-    #         "to": to_number,
-    #         "type": "template",
-    #         "template": {
-    #             "name": template_name,
-    #             "language": {
-    #                 "code": language
-    #             }
-    #         }
-    #     }
-
-    #     # Add components/parameters if provided
-    #     if params:
-    #         components = []
-    #         for param_type, values in params.items():
-    #             component = {"type": param_type, "parameters": []}
-    #             for value in values:
-    #                 component["parameters"].append(value)
-    #             components.append(component)
-
-    #         if components:
-    #             payload["template"]["components"] = components
-
-    #     try:
-    #         logger.info(f"Sendy WhatsApp template: {payload}")
-    #         response = requests.post(self.api_url, headers=headers, json=payload)
-    #         response.raise_for_status()
-    #         return response.json()
-    #     except requests.exceptions.RequestException as e:
-    #         logger.error(f"Failed to send WhatsApp template: {e}")
-    #         return {"error": str(e)}
-
-    def send_template(self, to_number: str, template_name: str, language: str = "en",
-                  params: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
-        """
-        Send a WhatsApp template message
-
-        Args:
-            to_number: Recipient's phone number with country code (e.g. "911234567890")
-            template_name: Name of the template to send
-            language: Language code for the template
-            params: Parameters to pass to the template (if any)
-
-        Returns:
-            API response as dictionary
-        """
-        logger.info(f"Sending WhatsApp template to {to_number} - {template_name}")
-        headers = {
-            "Authorization": f"Bearer {self.access_token}",
-            "Content-Type": "application/json"
-        }
-
-        # Basic payload for template message
-        payload = {
-            "messaging_product": "whatsapp",
-            "to": to_number,
-            "type": "template",
-            "template": {
-                "name": template_name,
-                "language": {
-                    "code": language
-                }
-            }
-        }
-
-        # Add components/parameters if provided
-        if params:
-            components = []
-            for component_type, values in params.items():
-                if component_type.lower() == "button":
-                    # Create a separate component for each button.
-                    for button in values:
-                        # Extract optional keys for button, defaulting if not provided.
-                        sub_type = button.pop("sub_type", "url")
-                        index = button.pop("index", 0)
-                        component = {
-                            "type": "button",
-                            "sub_type": sub_type,
-                            "index": index,
-                            "parameters": [button]
-                        }
-                        components.append(component)
-                else:
-                    # For other types like "body", group all provided parameters in one component.
-                    component = {
-                        "type": component_type,
-                        "parameters": values
-                    }
-                    components.append(component)
-            if components:
-                payload["template"]["components"] = components
-
-        try:
-            logger.info(f"Sendy WhatsApp template: {payload}")
-            response = requests.post(self.api_url, headers=headers, json=payload)
-            response.raise_for_status()
-            return response.json()
-        except requests.exceptions.RequestException as e:
-            logger.error(f"Failed to send WhatsApp template: {e}")
-            return {"error": str(e)}
-
-
-    def send_text_message(self, to_number: str, text: str) -> Dict[str, Any]:
-        """
-        Send a regular text message using WhatsApp API
-
-        Args:
-            to_number: Recipient's phone number with country code
-            text: Message text to send
-
-        Returns:
-            API response as dictionary
-        """
-        headers = {
-            "Authorization": f"Bearer {self.access_token}",
-            "Content-Type": "application/json"
-        }
-
-        payload = {
-            "messaging_product": "whatsapp",
-            "recipient_type": "individual",
-            "to": to_number,
-            "type": "text",
-            "text": {
-                "preview_url": False,
-                "body": text
-            }
-        }
-
-        try:
-            response = requests.post(self.api_url, headers=headers, json=payload)
-            response.raise_for_status()
-            return response.json()
-        except requests.exceptions.RequestException as e:
-            logger.error(f"Failed to send WhatsApp message: {e}")
-            return {"error": str(e)}
+    
+    def __init__(self, *args, **kwargs):
+        warnings.warn(
+            "Importing WhatsAppSender from care_im.utils.send_message_templates is deprecated. "
+            "Please import from care_im.messaging.template_sender instead.",
+            DeprecationWarning,
+            stacklevel=2
+        )
+        super().__init__(*args, **kwargs)
